Use a local copy of the stats structures in the reader and writer functions
The reader and writer functions used their argument as a pointer to a
stats structure, and wrote their stats to it directly in memory inside
the loop.  Use a local variables for stats structures instead and
write out the statistics after the loop; this allows the compiler to
keep the statistics in registers.